pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pinlite"
version = "0.1.0"
description = "Self-hosted media storage with CAS deduplication"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "loguru>=0.7.0",
    "python-multipart>=0.0.6",
    "python-magic>=0.4.27",
    "pillow>=10.1.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "types-aiofiles>=23.2.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["backend"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4"]
# B008: Do not perform function call in argument defaults (FastAPI Depends pattern)
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
files = ["backend", "tests"]
plugins = ["pydantic.mypy"]
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true

# Ослабляем правила для тестов (разрешаем неполные типы, но требуем return type)
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
